refactor: core presigned url api 반영
